upd: start refactoring
